Move Models unit tests to its own folder
