+++ tests/Unit/Models/ControlRuleTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit;
+namespace Tests\Unit\Models;
 
 use App\Models\ControlRule;
 use Tests\ModelTestCase;
