+++ tests/Unit/Models/KeygroupTest.php
@@ -1,10 +1,24 @@
 <?php
+/*
+ * SSH Access Manager - SSH keys management solution.
+ *
+ * Copyright (c) 2017 - 2021 by Paco Orozco <paco@pacoorozco.info>
+ *
+ *  This file is part of some open source application.
+ *
+ *  Licensed under GNU General Public License 3.0.
+ *  Some rights reserved. See LICENSE, AUTHORS.
+ *
+ *  @author      Paco Orozco <paco@pacoorozco.info>
+ *  @copyright   2017 - 2021 Paco Orozco
+ *  @license     GPL-3.0 <http://spdx.org/licenses/GPL-3.0>
+ *  @link        https://github.com/pacoorozco/ssham
+ */
 
-namespace Tests\Unit;
+namespace Tests\Unit\Models;
 
 use App\Models\Keygroup;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
-use Illuminate\Database\Eloquent\Relations\HasMany;
 use Tests\ModelTestCase;
 
 class KeygroupTest extends ModelTestCase
